Starts converting latlng tuple to literal
